nmc-35 update examples
