+++ examples/auth-intl/src/proxy.ts
@@ -1,7 +1,7 @@
 import { default as nextAuth } from "next-auth/middleware";
-import createMiddleware from "next-intl/middleware";
-import { chain } from "@nimpl/middleware-chain";
 import { NextResponse } from "next/server";
+import { chain } from "@nimpl/proxy-chain";
+import createMiddleware from "next-intl/middleware";
 
 const intlMiddleware = createMiddleware({
     locales: ["en", "dk"],
@@ -19,5 +19,5 @@ export default chain([
 ]);
 
 export const config = {
-    matcher: ["/((?!api|_next/static|_next/image|favicon.ico).*)"],
+    matcher: ["/((?!api|_next/static|_next/image|favicon.ico|.well-known).*)"],
 };
