+++ examples/auth5-intl/package.json
@@ -5,21 +5,20 @@
   "scripts": {
     "dev": "next dev",
     "build": "next build",
-    "start": "next start",
-    "lint": "next lint"
+    "start": "next start"
   },
   "dependencies": {
-    "@nimpl/middleware-chain": "0.5.2",
-    "next": "14.2.3",
+    "@nimpl/proxy-chain": "latest",
+    "next": "16.0.0",
     "next-auth": "5.0.0-beta.19",
     "next-intl": "3.14.1",
-    "react": "^18",
-    "react-dom": "^18"
+    "react": "19.2.0",
+    "react-dom": "19.2.0"
   },
   "devDependencies": {
-    "@types/node": "^20",
-    "@types/react": "^18",
-    "@types/react-dom": "^18",
-    "typescript": "^5"
+    "@types/node": "24.9.1",
+    "@types/react": "19.2.2",
+    "@types/react-dom": "19.2.2",
+    "typescript": "5.9.3"
   }
 }
