+++ examples/auth5-intl/src/auth.ts
@@ -1,4 +1,3 @@
-/* eslint-disable no-param-reassign */
 import NextAuth from "next-auth";
 import Google from "next-auth/providers/google";
