+++ examples/base/src/app/layout.tsx
@@ -1,4 +1,5 @@
-import type { Metadata } from "next";
+import { type Metadata } from "next/types";
+
 import "./globals.css";
 
 export const metadata: Metadata = {
